more network, pushing for suse
